Update timers to improve playability
It also removes the call to get_tree().paused when cleaning
completed rows as it is not needed to stop the gameplay, just
to stop the main timer to avoid the next shape to start falling
while completed rows are still being cleaned.

That call could also produce some weird behaviours as during that
pause some key pressed events could be missed.