Fix some warnings
Ignore return values when we don't need them. These sentences
should always be valid movements, in other case it would be a
bug. Addind if sentences would be double-checking that these
movements are valid, which seems redundant and would have no
effect on the game because if they fail, the gameplay will
be buggy which is the same as logging some warnings or something
similar.